+++ tests/test_api.py
@@ -303,3 +303,27 @@ def test_basic_2_wrong_order_1():
     assert results['additional_checks'][0]['referenced_from'] == 'interestedParty'
     assert results['additional_checks'][0]['entity_statement_out_of_order'] == '9bf27aa7-f372-41d7-9429-1bcd8b0f475d'
     assert results['additional_checks'][0]['seen_in_ownership_or_control_statement'] == 'fbfd0547-d0c6-4a00-b559-5c5e91c34f5c' # noqa
+
+
+def test_basic_bad_identifier_scheme():
+
+    cove_temp_folder = tempfile.mkdtemp(prefix='lib-cove-bods-tests-', dir=tempfile.gettempdir())
+    json_filename = os.path.join(os.path.dirname(
+        os.path.realpath(__file__)), 'fixtures', 'api', 'basic_bad_identifier_scheme.json'
+    )
+
+    results = bods_json_output(cove_temp_folder, json_filename)
+
+    assert results['validation_errors_count'] == 0
+    assert results['additional_fields_count'] == 0
+    assert results['additional_checks_count'] == 1
+    assert results['file_type'] == 'json'
+    assert results['statistics']['count_entity_statements'] == 1
+    assert results['statistics']['count_person_statements'] == 1
+    assert results['statistics']['count_ownership_or_control_statement'] == 1
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_person'] == 1
+    assert results['statistics']['count_ownership_or_control_statement_interested_party_with_entity'] == 0
+
+    assert results['additional_checks'][0]['type'] == 'entity_identifier_scheme_not_known'
+    assert results['additional_checks'][0]['scheme'] == 'GB-COH-THIS-SCHEME-IS-NOT-REAL-I-JUST-MADE-IT-UP-MWAHAHAHAHA'
+    assert results['additional_checks'][0]['entity_statement'] == '1dc0e987-5c57-4a1c-b3ad-61353b66a9b7'