+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Additional Checks: All relevant statements are present
 - Additional Checks: All relevant statements are in correct order
 - Additional Checks: Person or entity statements should be used in an ownership-or-control statement
+- Additional Checks: scheme should be an org-id code (This requires lib-cove 0.2.0 or higher)
 - Statistic: % of Interests of each type
 
 ### Changed